requirements/base.txt
django-cms==2.4.3
django-filer==0.9.5
cmsplugin-filer==0.9.5
